pyproject.toml
[project]
name = "luximus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiomysql>=0.2.0",
    "alembic>=1.14.1",
    "asgiref>=3.8.1",
    "celery>=5.4.0",
    "fastapi>=0.115.8",
    "google-api-python-client>=2.160.0",
    "google-auth>=2.38.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "honcho>=2.0.0",
    "letta-client>=0.1.24",
    "mysql-connector-python>=9.2.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.7.1",
    "pydantic[email]>=2.10.6",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "pytz>=2025.1",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.37",
    "taskipy>=1.14.1",
    "uvicorn>=0.34.0",
]

[tool.taskipy.tasks]
ngrok = "ngrok http 8000"
serve = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"
makemigration = "alembic revision --autogenerate -m"
migrate = "alembic upgrade head"
start = "honcho start"
reset-redis = "uv run scripts/reset_redis.py"
